cmemprof: add manual frame pointer unwinding, use uintptr_t for addresses
When hitting the C-Go boundary (at runtime.asmcgocall), call stack
unwinding stops. At this point we should be able to continue up the Go
call stack through frame pointer unwinding. Do this, currently only for
X86_64 (should be easy to replace the register later for different
architectures) to cross into Go.

As a result of this change, internal/testapp would panic with the error
"cgo argument has Go pointer to Go pointer" during the call to read a
sample. This is because the stack array was holding pointers to Go
functions. The CGo guide specifically calls out passing Go pointers to
structures containing other Go pointers to C as invalid (see
https://pkg.go.dev/cmd/cgo#hdr-Passing_pointers). To be more correct,
pass around the instruction pointers as uintptr_ts instead.