cmemprof: try out using runtime.Callers for stack traces
Basically try to do less custom stack unwinding. runtime.Callers should
handle the Go parts of the call stack appropriately, but will require
runtime.SetCgoTraceback functions to be set to handle the C parts of the
call. On the other hand, if somebody is using this profile plus CPU
profiling, they'll likely want to do that anyway.