cmemprof: work around crash when calling Go from new Go OS threads
When calling from C to Go, the Go runtime must find an "m" (essentially
an OS thread) and run the Go code on the corresponding "g" (goroutine)
running on the thread. This happens in runtime.cgocallback.  However,
when creating a new "m" there will be no "g" assigned to the "m". For
cgo programs, the function x_cgo_thread_create is responsible for
creating the new OS thread backed by a new "m". malloc is called to
store some thread-specific data, which goes through our malloc wrapper.
If goCallers happens to be called at this time then runtime.cgocallback
will crash by dereferencing a null pointer when accessing the
non-existent "g" of the newly-created "m".

To work around this crash, we use a combination of the GNU linker's
--wrap option and GCC's "weak" attribute to wrap x_cgo_thread_start
and set a flag to indicate we're in that function. Then in
profile_allocation we check the flag and do nothing if it's set.
It's not 100% clear to me _why_ this works. I copied the idea from
this Stack Overflow answer: https://stackoverflow.com/a/47094656

(This works on Linux, haven't tested on macos)