cmemprof: sample based on amount allocated rather than fixed rate
Copying the method of the Go blocking event profiler, given a rate R
(bytes) we sample an N byte allocation with probability min(1, N / R).
In practice this achieves a similar goal to the more common but also
more difficult to grok Poisson point process sampling employed in other
allocation profiles. The idea is that larger allocations, which might be
more important in a heap profile, are more likely to be sampled.
Frequent small allocations can also still be profiled.