internal/cgosymbolizer: add (vendored) cgosymbolizer fork
Vendored version of github.com/ianlancetaylor/cgosymbolizer with some
additions as part of a work-in-progress to see if it can be made
signal-safe and support the "context" callback. The dl_iterate_phdr
function (called as part of unwinding) is not signal-safe and so the
internal/testapp can deadlock if CPU profiling interrupts stack
unwinding to do its own unwinding. The workaround here is based on the
code found here:

  https://github.com/alk/unwind_safeness_helper/blob/master/unwind_safeness_helper.c

In summary, cut unwinding short if the current thread is already in
dl_iterate_phdr.

There are a few issues that need to be fixed:

* I copied the idea of making a thread-local context struct from
  github.com/benesch/cgosymbolizer. However, I don't think this is
  correct. A new context is created for each C -> Go transition.  As
  such, a goroutine can have several contexts. The contexts are
  basically appended to a sequence that's visited in reverse order when
  unwinding the stack. So presumably a single thread-local context would
  be clobbered by multiple C -> Go transitions in the same thread.

* The context callback currently just does the entire stack trace. This
  is undesirable because the context function will be called every time
  C code calls Go code, but the information from it is only needed if we
  unwind the call. So this would introduce a lot of overhead.

* The method of stopping unwinding based on whether the thread is
  already in dl_iterate_phdr means that some stack traces will be cut
  short/incomplete in the final profiles. Maybe that's not a big deal
  but we should see how often it happens in practice.