cmemprof: pull in dd-trace-go changes
This project started out as a research project. It moved into
dd-trace-go, and was planned to be included as a fully supported
extension. However, it's been harder than expected to justify continued
work on the project. It employs a few unpleasant workarounds for crashes
due to unusual interactions with the Go runtime. These workarounds
depend on implementation details that could change at any time. This
means it's hard to support this project. Also, the overhead is quite
high so it's unclear whether this is suitable for use in anything other
than throwaway testing situations.

Given all that, I'm pulling back the changes into this personal project
so I can keep hacking on it in my spare time.